Up to exercise 3.12 (IMPT: Read extra message)
'phonebook_backend' required a new git repository, and so this folder
was initialized with git init. It pushes to a Heroku repository that
allows the application to run on the service's servers.

Be careful with handling this FSOsubmissions repository in the future.
Remember to first commit the subrepository THEN this repository. See
git warning:

hint: You've added another git repository inside your current repository.
hint: Clones of the outer repository will not contain the contents of
hint: the embedded repository and will not know how to obtain it.
hint: If you meant to add a submodule, use:
hint:
hint:   git submodule add <url> part3/phonebook_backend
hint:
hint: If you added this path by mistake, you can remove it from the
hint: index with:
hint:
hint:   git rm --cached part3/phonebook_backend
hint:
hint: See "git help submodule" for more information.

Other changes: npm production builds were created in part2 to transfer
as static files to the pt3/backend. We can now run the backend solely
and we will have both a backend and a frontend. Proxies were added to
the frontend to enable development with the frontend (provided you also
run the backend localhost server).

The website is operational on the link provided in the README.md. It can
be updated in future exercises.